requirements.txt
fastapi
uvicorn[standard]
python-multipart
docling
docling-google-ocr
langgraph
langchain-core
pydantic
sqlalchemy
pymongo
motor
python-dotenv
google-generativeai
google-genai
aiohttp
aiofiles
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]
